changes and features
feat: Changed "cherrypy" to "hypercorn".
This was changed to be able to use asynchrony in the project.

chore: Reorganize project folder structure.

feat: Asynchronization was added to different functions that need it.
This is to reduce the response times of both the database and the response with the desired information.

refactor(users): The classes containing the controllers were changed to functions.

refactor(folders):  The classes containing the controllers were changed to functions.

refactor(events):  The classes containing the controllers were changed to functions.

refactor(tasks):  The classes containing the controllers were changed to functions.